+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.0] - 2024-07-31
+- added endpoint for health check
+- added cli command to activate all global regions
+
 ## [3.1.0] - 2024-05-24
 - added a new endpoint to change your password